Update failed to connect message for iOS
Prompt to forget micro:bit in settings. For the scenario where the app thinks it is bonded, whereas the micro:bit is not (e.g. full flash micro:bit using computer after doing a successfully flashing micro:bit with app), user needs to manually forget the micro:bit in settings.