Update failed to connect message for iOS (#9)
Add prompt to forget micro:bit in iOS settings. This is for the scenario where the app thinks it is bonded, whereas the micro:bit is not (e.g. full flash micro:bit using computer after doing a successfully flashing micro:bit with app). User needs to manually forget the micro:bit in settings to fix the issue.

“Try again” seems to work without needing to forget/unpair micro:bit for Android. We can potentially cope with this issue automatically for Android in the future.